+++ src/connection.h
@@ -38,6 +38,15 @@ extern "C" {
 #define SHUTDOWN_REASON_SIGTERM 		"SIGTERM"
 
 /**
+* parodusOnConnStatusChangeHandler - Function pointer
+* Used to define callback function to do additional processing
+* when websocket cloud connection status change event
+* i.e. "cloud-conn-status" as "fail" or "success"
+*/
+typedef void (*parodusOnConnStatusChangeHandler) (char * status);
+extern parodusOnConnStatusChangeHandler on_conn_status_change;
+
+/**
 * parodusOnPingStatusChangeHandler - Function pointer
 * Used to define callback function to do additional processing 
 * when websocket Ping status change event 
@@ -78,6 +87,9 @@ void set_cloud_disconnect_time(int disconnTime);
  */
 void start_conn_in_progress (unsigned long start_time);
 void stop_conn_in_progress (void);
+
+// To Register parodusOnConnStatusChangeHandler Callback function
+void registerParodusOnConnStatusChangeHandler(parodusOnConnStatusChangeHandler on_conn_status_change);
 
 // To Register parodusOnPingStatusChangeHandler Callback function
 void registerParodusOnPingStatusChangeHandler(parodusOnPingStatusChangeHandler on_ping_status_change);